Add basic gui versionizer, refactor code

+++ libUtils/src/main/java/de/thecodelabs/utils/io/IOUtils.java
@@ -9,7 +9,7 @@ public class IOUtils
 
 	public interface CopyDelegate
 	{
-		void complete(int bytes);
+		void complete(long bytes);
 	}
 
 	/**
